refactor: extract LookupPanel into dedicated component and clean up App styles
