+++ src/App/App.css
@@ -87,7 +87,9 @@
 }
 
 .App-StatValue {
-  display: block;
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
   font-size: 1.35rem;
   font-weight: 700;
   color: var(--color-text);
@@ -99,12 +101,6 @@
   font-size: 0.8rem;
   line-height: 1.4;
   color: var(--color-text-subtle);
-}
-
-.App-StatValue {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
 }
 
 .App-StatFlags {
@@ -119,163 +115,6 @@
   to { opacity: 1; transform: scale(1); }
 }
 
-.LookupPanel {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  max-width: 760px;
-  margin: 0 auto;
-  padding: 1rem;
-  border: 1px solid rgba(255, 255, 255, 0.08);
-  border-radius: 28px;
-  background:
-    linear-gradient(180deg, rgba(255, 255, 255, 0.06), rgba(255, 255, 255, 0.03)),
-    rgba(12, 14, 19, 0.88);
-  box-shadow:
-    0 24px 70px rgba(0, 0, 0, 0.42),
-    inset 0 1px 0 rgba(255, 255, 255, 0.05);
-  backdrop-filter: blur(28px);
-  animation: panelRise 0.75s ease-out;
-  max-height: calc(100dvh - 48px);
-}
-
-.App_active .LookupPanel {
-  flex: 1;
-}
-
-.LookupPanel-ResultsArea {
-  flex: 1;
-  overflow-y: auto;
-  margin-top: 1rem;
-  padding-right: 0.4rem;
-  padding-bottom: 2rem;
-}
-
-/* Hide scrollbar for Chrome, Safari and Opera */
-.LookupPanel-ResultsArea::-webkit-scrollbar {
-  display: none;
-}
-/* Hide scrollbar for IE, Edge and Firefox */
-.LookupPanel-ResultsArea {
-  -ms-overflow-style: none;  /* IE and Edge */
-  scrollbar-width: none;  /* Firefox */
-}
-
-.LookupPanel-Topbar {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 1rem;
-  padding: 0.15rem 0 0.85rem;
-}
-
-.LookupPanel-Dots {
-  display: flex;
-  gap: 0.45rem;
-}
-
-.LookupPanel-Dot {
-  width: 0.7rem;
-  height: 0.7rem;
-  border-radius: 999px;
-  background: rgba(255, 255, 255, 0.12);
-}
-
-.FlagToggle {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.4rem 0.85rem;
-  border: 1px solid rgba(255, 255, 255, 0.08);
-  border-radius: 999px;
-  background: rgba(255, 255, 255, 0.03);
-  color: var(--color-text-muted);
-  font-size: 0.75rem;
-  font-weight: 600;
-  letter-spacing: 0.02em;
-  text-transform: uppercase;
-  cursor: pointer;
-  transition: all 0.2s cubic-bezier(0.4, 0, 0.2, 1);
-}
-
-.FlagToggle:hover {
-  background: rgba(255, 255, 255, 0.06);
-  border-color: rgba(255, 255, 255, 0.15);
-  color: var(--color-text);
-}
-
-.FlagToggle_active {
-  background: rgba(255, 153, 56, 0.1);
-  border-color: rgba(255, 153, 56, 0.25);
-  color: #ff9938;
-}
-
-.FlagToggle-Icon {
-  font-size: 0.9rem;
-  filter: grayscale(1);
-  transition: filter 0.2s ease;
-}
-
-.FlagToggle_active .FlagToggle-Icon {
-  filter: grayscale(0);
-}
-
-.LookupPanel-Command {
-  padding: 1rem;
-  border-radius: 20px;
-  background: rgba(255, 255, 255, 0.03);
-  border: 1px solid rgba(255, 255, 255, 0.05);
-}
-
-.LookupPanel-CommandMeta {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 1rem;
-  margin-bottom: 0.85rem;
-}
-
-.LookupPanel-CommandLabel,
-.LookupPanel-CommandHint,
-.LookupPanel-SummaryLabel {
-  font-size: 0.8rem;
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-}
-
-.LookupPanel-CommandLabel {
-  color: var(--color-text-subtle);
-}
-
-.LookupPanel-CommandHint,
-.LookupPanel-SummaryLabel {
-  color: rgba(238, 240, 247, 0.45);
-}
-
-.LookupPanel-Summary {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 1rem;
-  padding: 1rem 0.4rem 0.7rem;
-}
-
-.LookupPanel-Tags {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.55rem;
-}
-
-.LookupPanel-Tags > span {
-  padding: 0.38rem 0.75rem;
-  border: 1px solid rgba(255, 255, 255, 0.08);
-  border-radius: 999px;
-  font-size: 0.84rem;
-  color: var(--color-text-muted);
-  background: rgba(255, 255, 255, 0.03);
-}
-
 @keyframes introRise {
   from {
     opacity: 0;
@@ -284,17 +123,6 @@
   to {
     opacity: 1;
     transform: translateY(0);
-  }
-}
-
-@keyframes panelRise {
-  from {
-    opacity: 0;
-    transform: translateY(20px) scale(0.985);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0) scale(1);
   }
 }
 
@@ -325,17 +153,5 @@
 
   .App-Stats {
     grid-template-columns: 1fr;
-  }
-
-  .LookupPanel {
-    scroll-snap-align: start;
-    scroll-margin-top: 24px;
-    min-height: calc(100dvh - 48px);
-  }
-
-  .LookupPanel-Summary,
-  .LookupPanel-CommandMeta {
-    flex-direction: column;
-    align-items: flex-start;
   }
 }
