+++ src/components/LanguageSwitcher/LanguageSwitcher.tsx
@@ -81,4 +81,4 @@ const LanguageSwitcher: React.FC = () => {
   );
 };
 
-export default LanguageSwitcher;
+export default React.memo(LanguageSwitcher);
